requirements.txt
pytest
requests
selenium
PyYAML
